Provided CLI Interface to CopyFill extract pattern
Provided CLI Interface To IsCopyNotFill extract Pattern

Default copy-fill pattern was 01 (fill, copy, fill copy etc...) This
allowed for the extraction of SPK documents. It has since been
discovered that the compression method used for ISPK documents is the
same but with a IsCopyNotFill pattern of 011 - with no other variations.
This new CLI argument allows the user to configure the extraction for
either an SPK or ISPK document. Currently, the only known useful
patterns are 01 (SPK) and 011 (ISPK)

Ie, this spk extractor now supports extraction of ISPK and SPK documents
via configuration of the copy-fill-pattern argument.

For purposes of simplicity, ISPK and SPK are considered the same file
format but with different extraction configurations.

Research is being done as to why there are multiple copy operations
side-by-side in ispk documents. It has been observed that copy
operations for skin, hair and armor of characters are separate (when
they could be merged into one.) Suggesting some pixel-to-material
mapping may be encoded in the compression algorithm. This meta-data
could be used for colouring character armor via the fragment shader.

Jeremy. Allen. Wildsmith.